backend/requirements.txt
torch
torchvision
pillow
numpy
typing
scikit-image
huggingface_hub>=0.20.0
# BriaRMBG (briaai/RMBG-1.4) requires 4.48.x; 4.49+ and 5.x use all_tied_weights_keys and break
transformers>=4.39.1
djangorestframework==3.15.2
Django==5.1.6
requests==2.32.3
gunicorn==23.0.0
django-cors-headers==4.7.0
whitenoise==6.5.0